Extract nodeProcessor from parser/index *
- Extract processExportDeclaration()
- Extract processExportKeyword()
