+++ src/parser/index.ts
@@ -9,23 +9,14 @@ import {
   TsConfig,
   TsConfigPaths,
 } from '../types';
-import { FromWhat, STAR } from './common';
-import { addDynamicImports, mayContainDynamicImports } from './dynamic';
-import {
-  addExportCore,
-  extractExport,
-  extractExportFromImport,
-  extractExportStatement,
-} from './export';
-import { addImportCore, extractImport } from './import';
 import { relative, resolve } from 'path';
 
-import { addImportsFromNamespace } from './imports-from-namespace';
+import { FromWhat } from './common';
+import { addExportCore } from './export';
+import { addImportCore } from './import';
 import { isNodeDisabledViaComment } from './comment';
+import { processNode } from './nodeProcessor';
 import { readFileSync } from 'fs';
-
-const hasModifier = (node: ts.Node, mod: ts.SyntaxKind): boolean | undefined =>
-  node.modifiers && node.modifiers.filter(m => m.kind === mod).length > 0;
 
 const extractFilename = (rootDir: string, path: string): string => {
   let name = relative(rootDir, path).replace(/([\\/]index)?\.[^.]*$/, '');
@@ -40,117 +31,6 @@ const extractFilename = (rootDir: string, path: string): string => {
   return name;
 };
 
-const processNode = (
-  node: ts.Node,
-  path: string,
-  addImport: (fw: FromWhat) => string | undefined,
-  addExport: (exportName: string, node: ts.Node) => void,
-  imports: Imports,
-  exports: string[],
-  extraOptions?: ExtraCommandLineOptions,
-  namespace = '',
-): void => {
-  const { kind } = node;
-
-  const processSubNode = (subNode: ts.Node, namespace: string): void => {
-    processNode(
-      subNode,
-      path,
-      addImport,
-      addExport,
-      imports,
-      exports,
-      extraOptions,
-      namespace,
-    );
-  };
-
-  if (kind === ts.SyntaxKind.ImportDeclaration) {
-    addImport(extractImport(node as ts.ImportDeclaration));
-    return;
-  }
-
-  if (kind === ts.SyntaxKind.ExportAssignment) {
-    addExport('default', node);
-    return;
-  }
-
-  if (kind === ts.SyntaxKind.ExportDeclaration) {
-    const exportDecl = node as ts.ExportDeclaration;
-    const { moduleSpecifier } = exportDecl;
-    if (moduleSpecifier === undefined) {
-      extractExportStatement(exportDecl).forEach(e => addExport(e, node));
-      return;
-    } else {
-      const { exported, imported } = extractExportFromImport(
-        exportDecl,
-        moduleSpecifier,
-      );
-      const key = addImport(imported);
-      if (key) {
-        const { what } = exported;
-        if (what == STAR) {
-          addExport(`*:${key}`, node);
-        } else {
-          what.forEach(w => exports.push(w));
-        }
-      }
-      return;
-    }
-  }
-
-  // Searching for dynamic imports requires inspecting statements in the file,
-  // so for performance should only be done when necessary.
-  if (mayContainDynamicImports(node)) {
-    addDynamicImports(node, addImport);
-  }
-
-  // Searching for use of types in namespace requires inspecting statements in the file,
-  // so for performance should only be done when necessary.
-  if (extraOptions?.searchNamespaces) {
-    addImportsFromNamespace(node, imports, addImport);
-  }
-
-  if (hasModifier(node, ts.SyntaxKind.ExportKeyword)) {
-    if (hasModifier(node, ts.SyntaxKind.DefaultKeyword)) {
-      addExport('default', node);
-      return;
-    }
-    const decl = node as ts.DeclarationStatement;
-    const name = decl.name ? decl.name.text : extractExport(path, node);
-
-    if (name) {
-      addExport(namespace + name, node);
-
-      if (extraOptions?.searchNamespaces) {
-        // performance: halves the time taken on large codebase (150k loc)
-        const isNamespace = node
-          .getChildren()
-          .some(c => c.kind === ts.SyntaxKind.NamespaceKeyword);
-
-        if (isNamespace) {
-          // Process the children, in case they *export* any types:
-          node
-            .getChildren()
-            .filter(c => c.kind === ts.SyntaxKind.Identifier)
-            .forEach(c => {
-              processSubNode(c, namespace + name + '.');
-            });
-
-          namespace = namespace + name + '.';
-        }
-      }
-    }
-  }
-
-  if (namespace.length > 0) {
-    // In namespace: need to process children, in case they *import* any types
-    node.getChildren().forEach(c => {
-      processSubNode(c, namespace);
-    });
-  }
-};
-
 const mapFile = (
   rootDir: string,
   path: string,
@@ -160,7 +40,7 @@ const mapFile = (
   extraOptions?: ExtraCommandLineOptions,
 ): File => {
   const imports: Imports = {};
-  const exports: string[] = [];
+  const exportNames: string[] = [];
   const exportLocations: LocationInFile[] = [];
   const name = extractFilename(rootDir, path);
 
@@ -181,7 +61,7 @@ const mapFile = (
   };
 
   const addExport = (exportName: string, node: ts.Node): void => {
-    addExportCore(exportName, file, node, exportLocations, exports);
+    addExportCore(exportName, file, node, exportLocations, exportNames);
   };
 
   ts.forEachChild(file, (node: ts.Node) => {
@@ -195,7 +75,7 @@ const mapFile = (
       addImport,
       addExport,
       imports,
-      exports,
+      exportNames,
       extraOptions,
     );
   });
@@ -204,7 +84,7 @@ const mapFile = (
     path: name,
     fullPath: path,
     imports,
-    exports,
+    exports: exportNames,
     exportLocations,
   };
 };
